+++ UI/src/TransitPanel.tsx
@@ -8,7 +8,7 @@ const transitLinesData$ = bindValue<string>("BetterTransitView", "transitLinesDa
 const showStopsAndStations$ = bindValue<boolean>("BetterTransitView", "showStopsAndStations", true);
 const showInfoviewBackground$ = bindValue<boolean>("BetterTransitView", "showInfoviewBackground", false);
 const showWaitingPassengers$ = bindValue<boolean>("BetterTransitView", "showWaitingPassengers", false);
-const showTransitVehicles$ = bindValue<boolean>("BetterTransitView", "showTransitVehicles", true);
+const showTransitVehicles$ = bindValue<boolean>("BetterTransitView", "showTransitVehicles", false);
 
 const VehicleIcon = memo(() => (<svg viewBox="0 0 24 24" style={{ width: '14rem', height: '14rem' }} fill="#bbb"><path d="M4 16c0 .88.39 1.67 1 2.22V20c0 .55.45 1 1 1h1c.55 0 1-.45 1-1v-1h8v1c0 .55.45 1 1 1h1c.55 0 1-.45 1-1v-1.78c.61-.55 1-1.34 1-2.22V6c0-3.5-3.58-4-8-4s-8 .5-8 4v10zm3.5 1c-.83 0-1.5-.67-1.5-1.5S6.67 14 7.5 14s1.5.67 1.5 1.5S8.33 17 7.5 17zm9 0c-.83 0-1.5-.67-1.5-1.5s.67-1.5 1.5-1.5 1.5.67 1.5 1.5-.67 1.5-1.5 1.5zm1.5-6H6V6h12v5z"/></svg>));
 const PassengerIcon = memo(() => (<svg viewBox="0 0 24 24" style={{ width: '14rem', height: '14rem' }} fill="#bbb"><path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z"/></svg>));
@@ -353,7 +353,7 @@ export const TransitPanel = () => {
         trigger("BetterTransitView", "setAllLinesVisible", targetState);
     };
 
-    const panelOpacity = showInfoviewBackground ? 1.0 : 0.95;
+    const panelOpacity = showInfoviewBackground ? 1.0 : 0.98;
 
     return (
         <div style={{